wallet/wallet.go
package wallet

import (
	"errors"
	"fmt"
)

type Bitcoin float64

type Wallet struct{
	balance Bitcoin
}

func (w *Wallet) Deposit(amount Bitcoin) {
	w.balance += amount
} 

var ErrInsufficientFunds = errors.New("Cannot withdraw, insufficient funds")

func (w *Wallet) Withdraw(amount Bitcoin) error {
	if (w.balance < amount) {
		return ErrInsufficientFunds
	}
	w.balance -= amount
	return nil
}

func (w *Wallet) Balance() Bitcoin {
	return w.balance
}

type Stringer interface {
	String() string
}

func (b Bitcoin) String() string {
	return fmt.Sprintf("%.2f BTC", b)
}
